% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQR_Momentum_Monthly.R
\docType{data}
\name{AQR.MOM.Monthly}
\alias{AQR.MOM.Monthly}
\title{AQR Momentum Indices, Monthly}
\format{
An xts object on Jan 1972 / Jul 2011 containing: Data: 475 entries, 22 total columns
\itemize{
 \item \strong{Date:} Date, monthly periodicity. Formatted in as a Date object in %Y-%m format.
 \item \strong{US.LC:} Excess returns of Momentum on Large-Cap and Mid-Cap U.S. Equities
 \item \strong{Us.SC:} Excess returns of Momentum on US Small-Cap U.S. Equities. 
 \item \strong{Int:} Excess returns of Momentum on International (Non-U.S.) Equities. 
}
}
\source{
The information has been provided to you solely for information purposes and
does not constitute an offer or solicitation of an offer, or any advice or recommendation,
to purchase any securities or other financial instruments, and may not be construed
as such. All information provided herein is impersonal and not tailored to
the needs of any person, entity, or group of persons. This information is
subject to further review and revision. AQR is not responsible for any errors
or omissions, regardless of the cause, for the results obtained from the use
of the information.

The information contained herein is only as current as of the date indicated,
and may be superseded by subsequent market events or for other reasons. Charts
and graphs provided herein are for illustrative purposes only. The information
has been developed internally and/or obtained from sources believed to be reliable;
however, AQR does not make any representation or warranty, express or implied,
as to the information’s accuracy, adequacy or completeness of such information,
nor does AQR recommend that the attached information serve as the basis of any
investment decision. Nothing contained herein constitutes investment, legal, tax
or other advice, nor is it to be relied on in making an investment or other decision.

There can be no assurance that an investment strategy will be successful.
Historic market trends are not reliable indicators of actual future market behavior
or future performance of any particular investment, which may differ materially,
and should not be relied upon as such. This information should not be viewed as
a current or past recommendation or a solicitation of an offer to buy or sell
any securities or to adopt any investment strategy.

Performance of all cited indices is calculated on a total-return basis with dividends reinvested.

The investment strategy and themes discussed herein may be unsuitable for
investors depending on their specific investment objectives and financial
situation. Please note that changes in the rate of exchange of a currency may
affect the value, price, or income of an investment adversely.

**Past performance of the Information is not an indication of future results.**

\url{https://www.aqr.com/Insights/Datasets/Value-and-Momentum-Everywhere-Original-Paper-Data}
}
\usage{
data('AQR.MOM.Monthly')
}
\description{
Momentum is an investment style that has been studied extensively by academics 
for many years. The evidence for momentum is pervasive and supported by almost
two decades of academic and practitioner research. Studies have documented 
momentum in the U.S. as far back as the Victorian age. The evidence also shows 
that momentum works broadly across asset classes, including foreign stocks, 
bonds, commodities, currencies, index futures, and global country selection. 
AQR has developed methodologies for U.S. and international markets that 
capture momentum in an intuitive and transparent way. The methodology can be 
applied to any universe of stocks. At present, AQR has created momentum indices 
for three distinct markets:
}
\details{
The AQR Momentum Index (Large-Cap and Mid-Cap U.S. Equities)
The AQR Small Cap Momentum Index (Small-Cap U.S. Equities)
The AQR International Momentum Index (Non-U.S. Equities)

Each of the indices is constructed using a simple and transparent methodology 
that captures the essence of the momentum effect:

The investment universe is screened based on market capitalization, liquidity 
and other considerations.

The stocks in the investment universe are ranked by their total return over 
the prior twelve months, excluding the last month.
The top 33% of stocks with the highest rank are included, and weighted based 
on their market capitalization.
The index is reconstituted quarterly.

Data loads lazily.
}
\section{Notes}{
 This file contains monthly returns of the test assets.
}

\examples{
 str(AQR.MOM.Monthly)
}
\keyword{datasets}
