% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Betting_Against_Beta_UMD_Monthly.R
\docType{data}
\name{BAB.Factors.UMD.Monthly}
\alias{BAB.Factors.UMD.Monthly}
\title{Betting Against Beta: Up Minus Down (UMD) factors, Monthly}
\format{
A xts object with 1121 observations on 29 variables:
\itemize{
 \item \strong{Date:} Date, monthly periodicity. Formatted in as a Date object in %Y-%m-%d format.
 \item \strong{EQ.AUS:} Returns, numerical. AUS equity porfolio that is 
 self-financing returns of equity Up Minus Down (UMD) factors. 
 Abbreviated from original name of "Equities ~ AUS"
 \item \strong{EQ.AUT:} Returns, numerical. AUT equity porfolio that is 
 self-financing returns of equity Up Minus Down (UMD) factors. 
 Abbreviated from original name of "Equities ~ AUT"
 \item \strong{EQ.BEL:} Returns, numerical. BEL equity porfolio that is 
 self-financing returns of equity Up Minus Down (UMD) factors. 
 Abbreviated from original name of "Equities ~ BEL"
 \item \strong{EQ.CAN:} Returns, numerical. CAN equity porfolio that is 
 self-financing returns of equity Up Minus Down (UMD) factors. 
 Abbreviated from original name of "Equities ~ CAN"
 \item \strong{EQ.CHE:} Returns, numerical. CHE equity porfolio that is 
 self-financing returns of equity Up Minus Down (UMD) factors. 
 Abbreviated from original name of "Equities ~ CHE"
 \item \strong{EQ.DEU:} Returns, numerical. DEU equity porfolio that is 
 self-financing returns of equity Up Minus Down (UMD) factors. 
 Abbreviated from original name of "Equities ~ DEU"
 \item \strong{EQ.DNK:} Returns, numerical. DNK equity porfolio that is 
 self-financing returns of equity Up Minus Down (UMD) factors. 
 Abbreviated from original name of "Equities ~ DNK"
 \item \strong{EQ.ESP:} Returns, numerical. ESP equity porfolio that is 
 self-financing returns of equity Up Minus Down (UMD) factors. 
 Abbreviated from original name of "Equities ~ ESP"
 \item \strong{EQ.FIN:} Returns, numerical. FIN equity porfolio that is 
 self-financing returns of equity Up Minus Down (UMD) factors. 
 Abbreviated from original name of "Equities ~ FIN"
 \item \strong{EQ.FRA:} Returns, numerical. FRA equity porfolio that is 
 self-financing returns of equity Up Minus Down (UMD) factors. 
 Abbreviated from original name of "Equities ~ FRA"
 \item \strong{EQ.GBR:} Returns, numerical. GBR equity porfolio that is 
 self-financing returns of equity Up Minus Down (UMD) factors. 
 Abbreviated from original name of "Equities ~ GBR"
 \item \strong{EQ.GRC:} Returns, numerical. GRC equity porfolio that is 
 self-financing returns of equity Up Minus Down (UMD) factors. 
 Abbreviated from original name of "Equities ~ GRC"
 \item \strong{EQ.HKG:} Returns, numerical. HKG equity porfolio that is 
 self-financing returns of equity Up Minus Down (UMD) factors. 
 Abbreviated from original name of "Equities ~ HKG"  
 \item \strong{EQ.IRL:} Returns, numerical. IRL equity porfolio that is 
 self-financing returns of equity Up Minus Down (UMD) factors. 
 Abbreviated from original name of "Equities ~ IRL"  
 \item \strong{EQ.ISR:} Returns, numerical. ISR equity porfolio that is 
 self-financing returns of equity Up Minus Down (UMD) factors. 
 Abbreviated from original name of "Equities ~ ISR"  
 \item \strong{EQ.ITA:} Returns, numerical. ITA equity porfolio that is 
 self-financing returns of equity Up Minus Down (UMD) factors. 
 Abbreviated from original name of "Equities ~ ITA"  
 \item \strong{EQ.JPN:} Returns, numerical. JPN equity porfolio that is 
 self-financing returns of equity Up Minus Down (UMD) factors. 
 Abbreviated from original name of "Equities ~ JPN"  
 \item \strong{EQ.NLD:} Returns, numerical. NLD equity porfolio that is 
 self-financing returns of equity Up Minus Down (UMD) factors. 
 Abbreviated from original name of "Equities ~ NLD"  
 \item \strong{EQ.NOR:} Returns, numerical. NOR equity porfolio that is 
 self-financing returns of equity Up Minus Down (UMD) factors. 
 Abbreviated from original name of "Equities ~ NOR"  
 \item \strong{EQ.NZL:} Returns, numerical. NZL equity porfolio that is 
 self-financing returns of equity Up Minus Down (UMD) factors. 
 Abbreviated from original name of "Equities ~ NZL" 
 \item \strong{EQ.PRT:} Returns, numerical. PRT equity porfolio that is 
 self-financing returns of equity Up Minus Down (UMD) factors. 
 Abbreviated from original name of "Equities ~ PRT" 
 \item \strong{EQ.SGP:} Returns, numerical. SGP equity porfolio that is 
 self-financing returns of equity Up Minus Down (UMD) factors. 
 Abbreviated from original name of "Equities ~ SGP" 
 \item \strong{EQ.SWE:} Returns, numerical. SWE equity porfolio that is 
 self-financing returns of equity Up Minus Down (UMD) factors. 
 Abbreviated from original name of "Equities ~ SWE" 
 \item \strong{EQ.USA:} Returns, numerical. USA equity porfolio that is 
 self-financing returns of equity Up Minus Down (UMD) factors. 
 Abbreviated from original name of "Equities ~ USA"
 \item \strong{AEP.GL:} Returns, numerical. Aggregate global equity porfolio 
 that is self-financing returns of equity Up Minus Down (UMD) factors. 
 Abbreviated from original name of "Aggregate Equity Portfolios ~ Global"
 \item \strong{AEP.GLUS:} Returns, numerical. Aggregate global equity porfolio 
 except USA that is self-financing returns of equity Up Minus Down (UMD) factors. 
 Abbreviated from original name of "Aggregate Equity Portfolios ~ Global Ex USA"
 \item \strong{AEP.EU:} Returns, numerical. Aggregate Europe equity porfolio 
 that is self-financing returns of equity Up Minus Down (UMD) factors. 
 Abbreviated from original name of "Aggregate Equity Portfolios ~ Europe"
 \item \strong{AEP.NA:} Returns, numerical. Aggregate North America equity porfolio 
 that is self-financing returns of equity Up Minus Down (UMD) factors. 
 Abbreviated from original name of "Aggregate Equity Portfolios ~ North America"
 \item \strong{AEP.PA:} Returns, numerical. Aggregate Pacific equity porfolio 
 that is self-financing returns of equity Up Minus Down (UMD) factors. 
 Abbreviated from original name of "Aggregate Equity Portfolios ~ Pacific"
}
}
\source{
Variables 
Size (SMB): total market value of equity (ME).
Value (HML): book equity (BE) divided by total market value of equity (ME). 
To obtain shareholders’ equity we use Stockholders’ Equity (SEQ) but if not 
available, we use the sum of Common Equity (CEQ) and Preferred Stocks (PSTK). 
If both SEQ and CEQ are unavailable, we proxy shareholders’ equity by Total 
Assets (AT) minus the sum of Total Liability (LT) and Minority Interest (MIB). 
To obtain book equity (BE), we subtract from shareholders’ equity the 
preferred stock value (PSTKRV, PSTKL or PSTK depending on availability). We 
assume that accounting variables are known with a minimum 6-month gap and 
align book price of the firm at the end of the firm’s fiscal year ending 
anywhere in calendar year t - 1 to June of calendar year t.
Momentum (UMD): return over the prior 12 months, skipping the most recent 
month.

Portfolios
The portfolio construction follows Fama and French (1992, 1993, and 1996) and 
Asness and Frazzini (2013) and Asness, Frazzini, and Pedersen (2013).
We form one set of portfolios in each country and compute global factor 
portfolios by weighting each country’s portfolio by the country’s total 
(lagged) market capitalization.
 
The market factor MKT is the value-weighted return on all available stocks 
minus the one-month Treasury bill rate.
 
The size, value, and momentum factors are constructed using six 
value-weighted portfolios formed on size and book-to-market and 1-year return.
At the end of each calendar month, stocks are assigned to two size-sorted 
portfolios based on their market capitalization.
For U.S. securities, the size breakpoint is the median NYSE market equity.
For International securities, the size breakpoint is the 80th percentile 
by country.
We use conditional sorts, first sorting on size, then on the second variable 
and rebalance portfolios monthly to maintain value weights.
Note that this approach is slightly different from Fama and French (1992, 
1993, and 1996) who use independent sorts. We prefer conditional sorts to 
ensure a balanced number of securities in each portfolio.
The size factor SMB is the average return on the 3 small portfolios minus 
the average return on the 3 big portfolios:
 
SMB =        1/3 (Small value + Small Neutral + Small Growth) 
             -1/3 (Big Value + Big Neutral + Big Growth)   

SMB portfolios are value-weighted. The size and book-to-market breakpoints 
are refreshed in June of each calendar year, and the portfolios are 
rebalanced every calendar month to maintain value weights.

The value factor HMLdevil follows Asness and Frazzini (2013). HMLdevil  is 
the average return on the two value portfolios minus the average return on 
the two growth portfolios:
HMLdevil = 1/2 (Small Value + Big Value) - 1/2 (Small Growth + Big Growth)

The superscript "devil" indicates that to compute book-to-market ratios we 
scale book equity (BE) by the current total market value of equity (ME) at 
the end of each month following Asness and Frazzini (2013).

HMLdevil portfolios are value-weighted. The size and book-to-market
breakpoints are refreshed every calendar month, and the portfolios are 
rebalanced every calendar month to maintain value weights.

The value factor HMLFF follows Fama and French (1992, 1993, and 1996). HMLFF 
is the average return on the two value portfolios minus the average return on 
the two growth portfolios:

HMLFF= 1/2 (Small Value + Big Value) - 1/2 (Small Growth + Big Growth)

The superscript "FF" indicates that to compute book-to-market ratios we scale 
book equity (BE) by the total market value of equity (ME) at the fiscal year 
end.
Note that for firms with fiscal year ending in December, this is the same 
measure as Fama and French (1992). For firms with fiscal year not ending in 
December, we use prices at the fiscal year-end date while Fama and French 
(1992) use December prices for all firms.
HMLFF portfolios are value-weighted. The size and book-to-market breakpoints 
are refreshed in June of each calendar year, and the portfolios are 
rebalanced every calendar month to maintain value weights.
The momentum factor UMD is the average return on the two high return 
portfolios minus the average return on the two low return portfolios:

UMD = 1/2 (Small Value + Big Value) - 1/2 (Small Growth + Big Growth)

UMD portfolios are value-weighted. The size and book-to-market breakpoints 
are refreshed every calendar month, and the portfolios are rebalanced every 
calendar month to maintain value weights.

References 

Asness, Cliff and Andrea Frazzini, 2013, The Devil in HML’s Detail, Journal 
of Portfolio Management, 39, 49-68.
Asness, Cliff, Andrea Frazzini and Lasse Pedersen, 2014, Quality Minus Junk, 
Working Paper, http://papers.ssrn.com/sol3/papers.cfm?abstract_id=2312432.
Fama, Eugene F. and Kenneth R. French, 1992, The Cross Section of Expected 
Stock Returns, Journal of Finance, 47, 427–465.
Fama, Eugene F. and Kenneth R. French, 1993, Common Risk Factors in the 
Returns on Stocks and Bonds, The Journal of Financial Economics, 33, 3-56.
Fama, Eugene F. and Kenneth R. French, 1996, Multifactor Explanations of 
Asset Pricing Anomalies, Journal of Finance, 51, 55–84.

Differences between AQR & Fama-French Factors
Sorting is a differentiator – We do a dependent sort (first sort by size and 
then by metric), whereas French does an independent sort.
Universe is a differentiator – We use the same data sources as French (CRSP 
and XF), and we apply filters to get close to French’s universe, but no 
guarantee that the universes will be exactly the same.

Historical refreshes are a differentiator – French refreshes his entire 
history every month, whereas AQR has not historically done that. We’ve just 
appended the last month to the time series. However as of late July, we 
started refreshing the historical data monthly.

Cutoffs are not a differentiator – We both use 30th and 70th NYSE percentiles 
for high/low and the median NYSE breakpoint for small/big.

Risk-free rate is not a differentiator – Both market and factor returns are 
in excess of the risk-free rate (t-bills).

Disclosures

The Information has been provided to you solely for information purposes and 
does not constitute an offer or solicitation of an offer, or any advice or 
recommendation, to purchase any securities or other financial instruments, 
and may not be construed as such. All Information provided herein is impersonal 
and not tailored to the needs of any person, entity or group of persons. This 
Information is subject to further review and revision. AQR is not responsible 
for any errors or omissions, regardless of the cause, for the results obtained 
from the use of the Information.

The Information contained herein is only as current as of the date indicated, 
and may be superseded by subsequent market events or for other reasons. Charts 
and graphs provided herein are for illustrative purposes only. The Information 
has been developed internally and/or obtained from sources believed to be 
reliable; however, AQR does not make any representation or warranty, express 
or implied, as to the Information’s accuracy, adequacy or completeness of such 
Information, nor does AQR recommend that the attached Information serve as the 
basis of any investment decision. Nothing contained herein constitutes 
investment, legal, tax or other advice, nor is it to be relied on in making an 
investment or other decision.

There can be no assurance that an investment strategy will be successful. 
Historic market trends are not reliable indicators of actual future market 
behavior or future performance of any particular investment which may differ 
materially, and should not be relied upon as such. This Information should 
not be viewed as a current or past recommendation or a solicitation of an offer 
to buy or sell any securities or to adopt any investment strategy.

Performance of all cited indices is calculated on a total-return basis with 
dividends reinvested.

The investment strategy and themes discussed herein may be unsuitable for 
investors depending on their specific investment objectives and financial 
situation. Please note that changes in the rate of exchange of a currency may 
affect the value, price or income of an investment adversely.

**Past performance of the Information is not an indication of future results.**

\url{https://www.aqr.com/Insights/Datasets/Betting-Against-Beta-Equity-Factors-Monthly}
}
\usage{
data('BAB.Factors.UMD.Monthly')
}
\description{
This file contains monthly self-financing returns of equity Up Minus Down 
(UMD) factors.
}
\details{
This data set is related to “Betting Against Beta” (Frazzini and Pedersen, 2014). 
A basic premise of the capital asset pricing model (CAPM) is that all agents 
invest in the portfolio with the highest Sharpe ratio, or expected excess return 
per unit of risk, and leverage or de-leverage this portfolio to suit their risk 
preferences. However, many investors — such as individuals, pension funds and 
mutual funds — are constrained in the leverage that they can take, and therefore 
overweight riskier securities instead. This behavior of tilting toward beta 
suggests high-beta assets require lower risk-adjusted returns than low-beta 
assets. Indeed, the security market line for U.S. stocks is too flat relative 
to the CAPM and is better explained by the CAPM with limited borrowing. This 
raises several questions: What is the magnitude of this anomaly relative to the 
size, value and momentum effects? Is betting against beta rewarded in other 
countries and asset classes? How does the return premium vary over time and in 
the cross section? How does one bet against beta? To explore these questions, 
we construct market-neutral betting-against-beta (BAB) factors, which are long 
leveraged low-beta assets and short high-beta assets.

This data set is an updated and extended version of the paper data, for which 
we provide long/short BAB equity factors. BAB equity factors are for U.S. 
equities and 23 international equity markets (as well as multiple international 
and global aggregates). Data is updated and extended monthly. We also provide 
the returns for several additional global factors for reference.

Data loads lazily.
}
\section{Notes}{
 This file contains daily self-financing excess returns of 
long/short equity Betting Against Beta (BAB) factors. 	
Copyright ©2014 Andrea Frazzini and Lasse Heje Pedersen
}

\examples{
 str(BAB.Factors.UMD.Monthly)
}
\keyword{datasets}
